extra
Support decl float & string
Support plus operation for string
Support Do While
Support div

+++ token.h
@@ -34,7 +34,7 @@ typedef struct {
   int lineNo, colNo;
   TokenType tokenType;
   int value;
-  double fvalue;
+  float fvalue;
 } Token;
 
 TokenType checkKeyword(char *string);
